refactor: move all providers into Provider.tsx
