--- src/app/providers.tsx
@@ -1,18 +1,0 @@
-"use client";
-import { QueryClientProvider } from "@tanstack/react-query";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import { Provider as ChakraProvider } from "@/components/ui/provider"
-import { getQueryClient } from "@/lib/get-query-client";
-
-export default function Providers({ children }: { children: React.ReactNode }) {
-  const queryClient = getQueryClient();
-
-  return (
-    <ChakraProvider>
-      <QueryClientProvider client={queryClient}>
-        {children}
-        <ReactQueryDevtools></ReactQueryDevtools>
-      </QueryClientProvider>
-    </ChakraProvider>
-  );
-}
